refactor(sonar): extract shared parseFileNamespacesArgs for conflicts/ghost CLI
